Demo - Compound Component Pattern
